backend/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
flake8==5.0.4
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.3.0
progress==1.6
psycopg2-binary==2.9.7
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.11.0
